+++ agrirouter-api-java-api/src/main/java/com/dke/data/agrirouter/api/service/messaging/cloud/OnboardingService.java
@@ -1,4 +1,4 @@
-package com.dke.data.agrirouter.api.service.onboard.cloud;
+package com.dke.data.agrirouter.api.service.messaging.cloud;
 
 import agrirouter.cloud.registration.CloudVirtualizedAppRegistration;
 import com.dke.data.agrirouter.api.dto.onboard.OnboardingResponse;
